refactor to use pom properly
