+++ reference_cegedim__automated_tests_implementation/README.md
@@ -27,7 +27,11 @@ reference_cegedim__automated_tests_implementation/
 ├── pages/                        # Page Object Model (Cegedim-specific)
 │   ├── base_page.py
 │   ├── login_page.py
-│   └── access_record_structured_page.py
+│   ├── home_page.py
+│   ├── nms_episode_page.py
+│   ├── gp_record_page.py
+│   ├── structured_record_page.py
+│   └── access_record_structured_page.py   # Compatibility facade
 └── tests/
     ├── conftest.py               # Test patient data (EMIS, TPP, Medicus)
     └── step_defs/                # Step definitions wiring Gherkin → Playwright
@@ -70,6 +74,12 @@ GP_CONNECT_PASSWORD=<your_password>
 > 4. Translate the confirmed selectors and flow into Python page object methods
 >
 > Never guess selectors or navigation paths — always verify with playwright-cli first.
+
+### Page Object Pattern (Required)
+
+- Implement page objects per real application screen, not per feature file.
+- Keep low-level interactions in screen-specific modules under `pages/`.
+- Keep feature-oriented facades (for example `access_record_structured_page.py`) thin and delegation-only to preserve step-definition compatibility.
 
 ## Required Delivery Workflow
