+++ reference_cegedim__automated_tests_implementation/pages/__init__.py
@@ -1,9 +1,17 @@
 from pages.base_page import BasePage
+from pages.home_page import HomePage
 from pages.login_page import LoginPage
+from pages.nms_episode_page import NmsEpisodePage
+from pages.gp_record_page import GpRecordPage
+from pages.structured_record_page import StructuredRecordPage
 from pages.access_record_structured_page import AccessRecordStructuredPage
 
 __all__ = [
     "BasePage",
+    "HomePage",
     "LoginPage",
+    "NmsEpisodePage",
+    "GpRecordPage",
+    "StructuredRecordPage",
     "AccessRecordStructuredPage",
 ]
